% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisyTimeline.R
\name{daisyTimeline}
\alias{daisyTimeline}
\title{Daisy Timeline}
\usage{
daisyTimeline(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame with columns 'date' and 'content', and optionally 'side'}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{elementId}{HTML element ID}
}
\description{
Create a Daisy Timeline
}
