% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisyTimeline.R
\name{daisyTimeline}
\alias{daisyTimeline}
\title{Daisy Timeline}
\usage{
daisyTimeline(data, date, title, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame}

\item{date}{Column containing date information (can be a formula)}

\item{title}{Column containing event labels (can be a formula)}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{elementId}{HTML element ID}
}
\description{
Create a Daisy Timeline
}
\examples{
# Create sample data
events_data <- data.frame(
  release_date = as.Date(c("2022-01-01", "2023-01-01")),
  package_name = c("dplyr", "ggplot2"),
  event = c("Major release", "Bug fixes")
)

# Using simple column names 
daisyTimeline(events_data, date = ~date, title = ~content)

# Create timeline with formula-style column references
daisyTimeline(events_data,
              date = ~format(as.Date(release_date), "\%Y"),
              title = ~event)

}
