% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisyTimeline.R
\name{daisyTimeline}
\alias{daisyTimeline}
\title{Daisy Timeline}
\usage{
daisyTimeline(data, date, event, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame}

\item{date}{Column containing date information}

\item{event}{Column containing event labels}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{elementId}{HTML element ID}
}
\description{
Create a Daisy Timeline
}
